Check DETERMINISTIC_VARY env to disable automatic vary
